Conflicts:
	src/man/Man.cpp
	src/man/corpus/Sensors.cpp
	src/man/include/Profiler.cpp
	src/man/vision/CMakeLists.txt
	src/man/vision/EdgeDetector.cpp
	src/man/vision/EdgeDetector.h
	src/man/vision/Field.h
	src/man/vision/FieldLines.cpp
	src/man/vision/FieldLinesDetector.cpp
	src/man/vision/HoughSpace.cpp
	src/man/vision/HoughSpace.h
	src/man/vision/Profiler.h
	src/man/vision/Threshold.cpp
	src/man/vision/Vision.cpp
	src/man/vision/Vision.h
	src/man/vision/VisualLine.cpp
	src/man/vision/VisualLine.h
	src/man/vision/offline/OfflineVision.cpp
	src/man/vision/test/EdgeDetectorTest.cpp
	src/man/vision/test/HoughSpaceTest.cpp
	src/tool/TOOL/Learning/VisionState.java
	src/tool/TOOL/Vision/TOOLVisionLink.cpp

+++ util/scripts/linux_setup.sh
@@ -5,23 +5,50 @@ if [ $# -ne 1 ]; then
 	exit 1
 fi
 
-echo "Downloading awesome free stuff! Also accept the Sun Java licence by \
-pressing TAB!"
-
-PACKAGES="build-essential cmake git-core sun-java6-jdk \
-python2.6-dev emacs cmake-curses-gui ccache curl aptitude"
+PACKAGES="build-essential cmake git-core \
+python2.7-dev emacs cmake-curses-gui ccache curl aptitude \
+ant qt4-dev-tools"
 
 echo "Are you on 64-bit linux?(y/n)"
 read IS64BIT
+echo "What version of Ubuntu are you on? (example: 11.04)"
+read VERSION
 
 if [ $IS64BIT == 'y' ]; then
   PACKAGES="$PACKAGES g++-4.4-multilib"
 fi
 
-sudo add-apt-repository ppa:sun-java-community-team/sun-java6
-sudo apt-get update
+if [ $VERSION == '10.10' ]; then
+    echo "Downloading Java. Accept the license by pressing TAB!"
+    sudo add-apt-repository ppa:sun-java-community-team/sun-java6
+    sudo apt-get update
+    sudo apt-get install sun-java6-jdk
+    sudo update-java-alternatives -s java-6-sun
+elif [ $VERSION == '11.04' ]; then
+    echo "Downloading Java. Accept the license by pressing TAB!"
+    sudo add-apt-repository ppa:ferramroberto/java
+    sudo apt-get update
+    sudo apt-get install sun-java6-jdk
+elif [ $VERSION == '11.10' ]; then
+    echo "Downloading Java. Accept the license by pressing TAB!"
+    sudo add-apt-repository ppa:ferramroberto/java
+    sudo apt-get update
+    sudo apt-get install sun-java6-jdk
+else
+    echo "That version is not supported. Please use 10.10 or 11.04. However, \
+if you are very sure of what you're doing, you can \
+finish running the script and install Sun Java manually when it is done."
+    echo ""
+    echo "Continue? (y/n)"
+    read CONTINUE
+    if [ $CONTINUE == 'n' ]; then
+	echo "Exiting."
+	exit 1
+    fi
+fi
+
+echo "Downloading awesome free stuff!"
 sudo apt-get install $PACKAGES
-sudo update-java-alternatives -s java-6-sun
 
 echo "Downloading and unpacking NBites files"
 
@@ -30,11 +57,23 @@ robocup=robocup.bowdoin.edu:/mnt/research/robocup
 nbites_dir=$PWD/../..
 lib_dir=$nbites_dir/lib
 
-naoqi=naoqi-sdk-$naoqi_version-linux-nbites.tar.gz
-naoqi_robocup=$robocup/software/nao/NaoQi/$naoqi_version/$naoqi
-naoqi_local=$lib_dir/naoqi-sdk-$naoqi_version-linux
+naoqi=naoqi-sdk-$naoqi_version-linux32.tar.gz
+atom=nbites-atom-toolchain-$naoqi_version.tar.gz
+geode=nbites-geode-toolchain-$naoqi_version.tar.gz
 
-ext=ext-nbites-linux.tar.gz
+naoqi_robocup=$robocup/software/nao/NaoQi/$naoqi_version/$naoqi
+atom_robocup=$robocup/software/nao/NaoQi/$naoqi_version/$atom
+geode_robocup=$robocup/software/nao/NaoQi/$naoqi_version/$geode
+
+naoqi_local=$lib_dir/naoqi-sdk-$naoqi_version-linux32
+atom_local=$lib_dir/atomtoolchain
+geode_local=$lib_dir/geodetoolchain
+
+if [ $IS64BIT == 'y' ]; then
+    ext=ext-nbites-linux64.tar.gz
+else
+    ext=ext-nbites-linux32.tar.gz
+fi
 ext_robocup=$robocup/software/$ext
 
 echo ""
@@ -45,11 +84,23 @@ echo "Downloading NaoQi"
 mkdir -p $lib_dir
 rsync -v $USER_NAME@$naoqi_robocup $lib_dir/
 
+echo "Downloading Atom toolchain"
+rsync -v $USER_NAME@$atom_robocup $lib_dir/
+
+echo "Downloading Geode toolchain"
+rsync -v $USER_NAME@$geode_robocup $lib_dir/
+
 echo "Unpacking NaoQi"
 
 pushd $lib_dir
-tar -xvzf $naoqi
+tar -xzf $naoqi
 rm $naoqi
+
+mkdir $atom_local
+tar -xzf $atom -C $atom_local --strip-components 1
+
+mkdir $geode_local
+tar -xzf $geode -C $geode_local --strip-components 1
 popd
 
 if [ $IS64BIT == 'y' ]; then
@@ -58,12 +109,12 @@ ctc=linux-x64-crosstoolchain.tar.gz
 ctc_robocup=$robocup/software/nao/cross_compiler_stuff/$ctc
 
 echo "Downloading the CTC"
-rsync -v $USER@$ctc_robocup $naoqi_local/
+rsync -v $USER_NAME@$ctc_robocup $naoqi_local/
 
 echo "Unpacking the CTC"
 
 pushd $naoqi_local
-tar -xvzf $ctc
+tar -xzf $ctc
 rm $ctc
 popd
 fi
@@ -74,16 +125,34 @@ rsync -v $USER_NAME@$ext_robocup $nbites_dir/
 
 echo "Unpacking ext"
 pushd $nbites_dir
-tar -xvzf $ext
+tar -xzf $ext
 rm $ext
 popd
 
+echo "Setting up git stuff .."
+./git_setup.sh
+
 echo "Setting up bash stuff ..."
 
-echo "export NBITES_DIR=$nbites_dir" >> nbites.bash
-echo "export AL_DIR=$naoqi_local" >> nbites.bash
-echo "export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$nbites_dir/ext/lib" >> ~/nbites.bash
-echo "export PATH=$PATH:$nbites_dir/ext/bin" >> ~/nbites.bash
-echo "Done! The last step is just to add the following line"
-echo "source $PWD/nbites.bash"
+nbites_bash=$nbites_dir/util/scripts/nbites.bash
+
+echo "export NBITES_DIR=$nbites_dir" >> $nbites_bash
+echo "export AL_DIR=$naoqi_local" >> $nbites_bash
+echo "export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$nbites_dir/ext/lib" >> $nbites_bash
+echo "export PATH=$nbites_dir/ext/bin:$PATH" >> $nbites_bash
+echo "Done! The last step is just to add the following line:"
+echo "source $nbites_bash"
 echo "to your .bashrc (which is in your home directory)"
+
+echo "Would you like this to be done automatically? (y/n)"
+read AUTO
+
+if [ $AUTO == 'y' ]; then
+    echo "#added by linux-setup.sh for RoboCup purposes" >> ~/.bashrc
+    echo "source $nbites_bash" >> ~/.bashrc
+    echo "You're good to go!"
+else
+    echo "Add the line manually, and you'll be all set up!"
+fi
+
+echo "NOW RESTART THE TERMINAL SO THAT .bashrc GETS RUN!"
